# platform/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Video-Challenge-Plattform API
  version: 0.1.0
  description: >-
    Öffentliche und authentifizierte Endpoints der Plattform. Authentifizierung per
    Bearer-Token (im Mock-Auth-Setup ist der Token die User-ID; `admin:<uuid>` für
    Admin-Rechte). Beträge in Cent, Zeitstempel in UTC, IDs als UUID.
    Handgepflegt — bei Endpunkt-Änderungen mit aktualisieren.
servers:
  - url: http://localhost:8080
    description: Lokal
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: CHALLENGE_FULL }
            message: { type: string }
    ChallengeSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        category: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ChallengeStatus' }
        selectionMode: { type: string, enum: [CREATOR_DECIDES, COMMUNITY_VOTE] }
        prizeAmountCents: { type: integer }
        maxSlots: { type: integer, example: 10 }
        submissionDeadline: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    ChallengeStatus:
      type: string
      enum:
        [DRAFT, PENDING_FUNDING, OPEN, FULL, SUBMISSIONS_CLOSED, IN_REVIEW,
         SELECTION, WINNER_LOCKED, PAID_OUT, CANCELLED, EXPIRED]
...
  responses:
    Unauthorized:
      description: Bearer-Token fehlt oder ungültig
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ErrorResponse:
      description: Fehler (Code + deutsche Meldung)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
paths:
...
  /v1/users:
    post:
      tags: [Users]
      summary: Registrierung (18+-Gate); ID dient als Bearer-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { isAdult: { type: boolean } }
              required: [isAdult]
      responses:
        '201':
          description: Erstellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  isAdult: { type: boolean }
        '403': { $ref: '#/components/responses/ErrorResponse' }
  /v1/users/me:
    get:
      tags: [Users]
      summary: Eigenes Profil
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Profil }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Users]
      summary: Eigenes Profil aktualisieren (Nutzername/Anzeigename/Bio)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string, description: "a–z 0–9 _, 3–20 Zeichen" }
                displayName: { type: string, maxLength: 40 }
                bio: { type: string, maxLength: 200 }
      responses:
        '200': { description: Aktualisiertes Profil }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /v1/users/me/challenges:
    get:
      tags: [Users]
      summary: Eigene erstellte + beigetretene Challenges
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Listen
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: array, items: { $ref: '#/components/schemas/ChallengeSummary' } }
                  joined:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ChallengeSummary'
                        - type: object
                          properties: { slotStatus: { type: string } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1/users/me/notifications:
    get:
      tags: [Users]
      summary: In-App-Mitteilungen des Nutzers (neueste zuerst)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Mitteilungen + Ungelesen-Zähler
          content:
            application/json:
              schema:
                type: object
                properties:
                  unreadCount: { type: integer }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        type: { type: string }
                        challengeId: { type: string, format: uuid, nullable: true }
                        title: { type: string }
                        body: { type: string }
                        read: { type: boolean }
                        createdAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/users/me/notifications/read:
    post:
      tags: [Users]
      summary: Alle Mitteilungen als gelesen markieren
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Anzahl markierter Mitteilungen
          content:
            application/json:
              schema:
                type: object
                properties: { marked: { type: integer } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...